pybind/mgr/prometheus: Move Metric class into its own module
